\name{mergeExprs2}
\alias{mergeExprs2}
\title{
Function to merge ExpressionSet object
}
\description{
Function \code{mergeExprs} from library \code{MergeMaid} has been modiffied to have to arguments: list of ExpressionSet objects and vector of datasets names. 
}
\usage{
mergeExprs2(arg, names)
}
\arguments{
  \item{arg}{ List of ExpressionSet objects}
  \item{names}{ Vector of datasets names}
}

\value{
A mergeExpressionSet object.
}
\author{
Ivana Ihnatova 
}

\keyword{ manip }

