\name{GEDM<--methods}
\docType{methods}
\alias{GEDM<--methods}
\alias{GEDM<-,MetaArray-method}
\alias{clinical<--methods}
\alias{clinical<-,MetaArray-method}
\alias{datanames<--methods}
\alias{datanames<-,MetaArray-method}

\title{Replacement Methods for \code{MetaArray} object}
\description{
Each of the methods replaces one slot of an object derived from \code{MetaArray} class
}
\section{Methods}{
\describe{

\item{\code{signature(object = "MetaArray")}}{
Method replaces one slot of an object derived from \code{MetaArray} class, e.g. \code{"GEDM<-"} replaces the \code{GEDM} slot etc. 
}
}}
\keyword{methods}