\name{plotpattern}
\alias{plotpattern}

\title{
Function to do plot signifficance of Meta-Analysis Patterns
}
\description{
Function plots signifficance of Meta-Analysis Patterns
}
\usage{
plotpattern(intx, method)
}
\arguments{
  \item{intx}{
A data frame with rows refering to Meta-Analysis Patterns and columns (from 5th to 8th) to signifficance of observed  pattern matches
}
  \item{method}{
Either number 1 or 2, otherwise no plot is provided. If 1 a line plot is made. If 2 a form of scatterplot is drawn. 
}
}
\references{Yang, X., Bentink, S. a Spang, R. 2005, Detecting Common Gene Expression Patterns in Multiple Cancer Outcome Entities, Biomedical Microdevices}
\author{
Ivana Ihnatova
}

\keyword{ hplot }

